generator-shared
