+++ crates/generator-macros/src/lib.rs
