+++ crates/generator-macros/src/partial_unchecked.rs
