+++ crates/generator-shared/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "prisma-client-rust-generator-utils"
+name = "prisma-client-rust-generator-shared"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,3 +8,5 @@ convert_case.workspace = true
 proc-macro2 = "1.0.39"
 quote = "1.0.18"
 syn = { version = "1.0.91", features = ["extra-traits"] }
+
+prisma-client-rust-sdk = { path = "../sdk" }
