+++ crates/generator/src/models/filter.rs
@@ -1,7 +1,5 @@
-use prisma_client_rust_sdk::prisma::prisma_models::{
-    walkers::{ModelWalker, RefinedFieldWalker},
-    FieldArity,
-};
+use prisma_client_rust_generator_shared::select_include::SelectableFields;
+use prisma_client_rust_sdk::prisma::prisma_models::walkers::ModelWalker;
 
 use crate::prelude::*;
 
@@ -10,31 +8,13 @@ pub fn r#macro(model: ModelWalker, module_path: &TokenStream) -> TokenStream {
 
     let name = format_ident!("_{}_filter", model.name().to_case(Case::Snake, true));
 
-    let fields = model.fields().map(|field| {
-        let field_name_snake = snake_ident(field.name());
-
-        let variant = match field.refine() {
-            RefinedFieldWalker::Scalar(_) => quote!(Scalar),
-            RefinedFieldWalker::Relation(relation_field) => {
-                let related_model_name_snake = snake_ident(relation_field.related_model().name());
-
-                let relation_arity = match &field.ast_field().arity {
-                    FieldArity::List => quote!(Many),
-                    _ => quote!(One),
-                };
-
-                quote!(Relation(#module_path #related_model_name_snake, #relation_arity))
-            }
-        };
-
-        quote!((#field_name_snake, #variant))
-    });
+    let selectable_fields = SelectableFields::new(model.fields(), module_path);
 
     quote! {
         ::prisma_client_rust::macros::filter_factory!(
             #name,
             #module_path #model_name_snake,
-            [#(#fields),*]
+            #selectable_fields
         );
     }
 }
